feat: remove carbon/styles
Now that we are using web components, we don't need to duplicate carbon styles
