+++ packages/ai-chat/package.json
@@ -96,22 +96,18 @@
   "peerDependencies": {
     "@carbon/icons": ">=11.53.0 <12.0.0",
     "@carbon/react": ">=1.68.0 <2.0.0",
-    "@carbon/styles": ">=1.68.0 <2.0.0",
-    "@carbon/themes": ">=11.53.0 <12.0.0",
     "@carbon/web-components": ">=2.37.0 <3.0.0",
     "@floating-ui/react": ">=0.25.0 <1.0.0",
-    "classnames": ">=2.5.0 <3.0.0",
     "lit": ">=3.1.0 <4.0.0",
-    "lodash-es": ">=4.17.0 <5.0.0",
     "react": ">=17.0.0 <20.0.0",
     "react-dom": ">=17.0.0 <20.0.0",
     "react-redux": ">=8.0.0 <10.0.0",
-    "redux": ">=4.0.0 <6.0.0",
-    "tabbable": ">=6.2.0 <7.0.0"
+    "redux": ">=4.0.0 <6.0.0"
   },
   "dependencies": {
     "@ibm/telemetry-js": "^1.9.1",
     "@lit/react": "^1.0.6",
+    "classnames": "^2.5.0",
     "color": "^4.2.3",
     "compute-scroll-into-view": "^3.1.0",
     "csv-stringify": "^6.5.2",
@@ -121,11 +117,13 @@
     "highlight.js": "^11.9.0",
     "intl-messageformat": "^10.7.15",
     "intl-pluralrules": "^2.0.1",
+    "lodash-es": "^4.17.0",
     "markdown-it": "^14.1.0",
     "markdown-it-attrs": "^4.1.6",
     "react-intl": "^7.1.6",
     "react-player": "2.15.1",
-    "swiper": "^11.1.1"
+    "swiper": "^11.1.1",
+    "tabbable": "^6.2.0"
   },
   "scripts": {
     "build": "npm run build:chat && npm run build:typedoc",
