+++ packages/ai-chat/src/chat/shared/styles/carbon.scss
@@ -8,6 +8,7 @@
 @use "sass:meta";
 @use "@carbon/styles/scss/themes";
 @use "@carbon/styles/scss/theme";
+@use "@carbon/styles/scss/reset";
 
 .cds--white {
   @include theme.theme(themes.$white);
