+++ packages/ai-chat/src/chat/shared/styles/styles.scss
@@ -7,7 +7,6 @@
 
 @use "../../styles/cssVariables";
 @use "./carbon";
-@use "@carbon/styles";
 @use "./animations";
 @use "./common";
 @use "../Chat";
